+++ CHANGELOG.md
@@ -6,7 +6,12 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased](https://github.com/thebigmunch/google-music-proto/tree/master)
 
-[Commits](https://github.com/thebigmunch/google-music-proto/compare/2.1.2...master)
+[Commits](https://github.com/thebigmunch/google-music-proto/compare/2.2.0...master)
+
+
+## [2.2.0](https://github.com/thebigmunch/google-music-proto/releases/tag/2.2.0) (2019-01-15)
+
+[Commits](https://github.com/thebigmunch/google-music-proto/compare/2.1.2...2.2.0)
 
 ### Added